fix(ui): align Section props with legacy Stripe; add Hero/Content aliases
