+++ src/components/sections/index.ts
@@ -1,0 +1,5 @@
+export { Section } from "./Section";
+export type { SectionProps } from "./Section";
+
+// Aliases so app code can adopt "Section" language without moving files yet.
+export { HeroStripe as HeroSection, ContentStripe as ContentSection } from "../stripes";
